api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "myapi"
version = "0.1.0"
description = "FastAPI + Mongo (Motor) API"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn[standard]==0.35.0",
  "pydantic[email]==2.11.7",
  "pydantic-settings==2.10.1",
  "motor==3.7.1",
  "python-jose[cryptography]==3.5.0",
  "passlib[bcrypt]==1.7.4",
  "bcrypt==4.3.0",
  "python-multipart==0.0.20"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*", "docs*"]
